kisan_help/requirements.txt
django
gTTS
scikit-learn
numpy
joblib
google-generativeai
python-dotenv
openai
